+++ contracts/light/CMTAT_LIGHT_BEACON_FACTORY.sol
@@ -83,7 +83,7 @@ contract CMTAT_LIGHT_BEACON_FACTORY is CMTATBeaconFactoryBase, ReentrancyGuard {
     * @return initializerData Encoded call to the CMTAT Light initializer.
     */
     function _initializerData(
-        CMTAT_LIGHT_ARGUMENT calldata cmtatArgument) internal pure returns(bytes memory initializerData) {
+        CMTAT_LIGHT_ARGUMENT calldata cmtatArgument) internal pure virtual returns(bytes memory initializerData) {
         initializerData = abi.encodeWithSelector(
             CMTATUpgradeableLight(address(0)).initialize.selector,
             cmtatArgument.CMTATAdmin,