+++ contracts/standard/CMTAT_TP_FACTORY.sol
@@ -91,7 +91,7 @@ contract CMTAT_TP_FACTORY is CMTATTransparentFactoryBase, ReentrancyGuard {
     */
      function _initializerData(
         // CMTAT function initialize
-        CMTAT_ARGUMENT calldata cmtatArgument) internal pure returns(bytes memory initializerData) {
+        CMTAT_ARGUMENT calldata cmtatArgument) internal pure virtual returns(bytes memory initializerData) {
         initializerData = abi.encodeWithSelector(
             CMTATStandardUpgradeable(address(0)).initialize.selector,
                   cmtatArgument.CMTATAdmin,